Tracking is using SH images now.

+++ Plugins/org.mitk.gui.qt.diffusionimaging.reconstruction/src/internal/QmitkQBallReconstructionView.h
@@ -105,7 +105,7 @@ protected slots:
    * @brief PreviewThreshold Generates a preview of the values that are cut off by the thresholds
    * @param threshold
    */
-  void PreviewThreshold(int threshold);
+  void PreviewThreshold(short threshold);
 
 protected:
 
@@ -115,10 +115,10 @@ protected:
   Ui::QmitkQBallReconstructionViewControls* m_Controls;
 
   template<int L>
-  void TemplatedAnalyticalQBallReconstruction(mitk::DataNode* dataNodePointer, float lambda, int normalization);
+  void TemplatedAnalyticalQBallReconstruction(mitk::DataNode* dataNodePointer, double lambda, int normalization);
 
   template<int L>
-  void TemplatedMultiQBallReconstruction(float lambda, mitk::DataNode*);
+  void TemplatedMultiQBallReconstruction(double lambda, mitk::DataNode*);
 
 private:
